Moves getEvents() action creator to eventActions.js
